remove obsolete HeroSection variants and tweak title position
