+++ src/layout/section/heroSection/index.tsx
@@ -1,1 +1,2 @@
-export * from './HeroSection'
+export { HeroSection, HeroSectionProps } from './HeroSection'
+export { HeroSection as default } from './HeroSection'
